rl_racing/requirement.txt
opencv-python==4.5.1.48
mss==6.1.0
pynput==1.7.3
Pillow==8.1.0
torch==1.7.1
matplotlib==3.3.4